pineapple: Rewrite frida script to avoid constants
Both Needle and spotifya-analyze (v1 by @plietar) were based on
signature scanning and finding "offsets" of the shannon encryption
functions. This was semi-reliable across versions however as months go
on these are breaking. Pineapple fixes this by not requiring detouring
the encryption functions however the server key must still be replaced
to trick the client into verifying us. In Needle this was done by
parsing the executable or library on the host then sending the offset to
the script using Frida's RPC feature. In Pineapple this parsing is now
done on the device with some additional checks to increase performance.
If this is too slow the parsing code could be rewritten as a CModule or
compiled on the host as a dynamic library (.so, .dylib, .dll) and
injected using Frida.